[LOD-197] feat: 유저 DB 정보 변경에 반영, Profile 생성 이전 및 초기화 API 생성

+++ src/main/java/net/lodgames/relation/block/vo/UserBlockListVo.java
@@ -15,5 +15,5 @@ public class UserBlockListVo {
     // 프로필
     private String nickname;
     private String image;
-    private Short basicImageIdx;
+    private Integer basicImageIdx;
 }
